+++ Sandbox/Examples/Testing/ECSTestingLayer.h
@@ -7,7 +7,7 @@ class ECSTestingLayer :public Layer
 public:
 	ECSTestingLayer();
 	void OnCreate()override;
-	void OnUpdate(double DeltaTime)override;
+	void OnUpdate(float DeltaTime)override;
 	void OnDestroy()override;
 	void OnGUI()override;
 private:
