+++ Sandbox/Examples/Testing/TestingLayer.h
@@ -4,7 +4,7 @@ class TestingLayer :public Layer {
 public:
     TestingLayer();
     void OnCreate()override;
-    void OnUpdate(double deltatime)override;
+    void OnUpdate(float deltatime)override;
     void OnGUI()override;
     void OnDestroy()override;
